editing model

+++ app/models/items.py
@@ -10,8 +10,8 @@ class Item(db.Model):
     user_id = db.Column(db.Integer,db.ForeignKey("users.id", ondelete="CASCADE"), primary_key=True)
     title = db.Column(db.String(255), nullable=False)
     description = db.Column(db.String(2000), nullable=False)
-    price = db.Column(db.Integer, nullable=False)
-   
+    price = db.Column(db.Float, nullable=False)
+
     def to_dict(self):
         return {
             'id': self.id,
